refactor: BaseEntity 패키지 이동
- 여러 도메인에서 쓰이므로 common 하위로
